==> analyses/01-wrangle-data.Rmd <==
---
title: "Wrangle the data into datasets for analyses"
output: html_document
---

## Preparation

```{r}
# Load libraries
library(tidyverse)
library(geiger)
library(here)
```

```{r}
# Functions
# Convert SD to SE
sd2se <- function(x, n) {
  x / sqrt(n)
}

# Convert SE to SD
se2sd <- function(x, n) {
  x * sqrt(n)
}

# Create the max column for each variable
get_max <- function(var.max, var.mean) {
  ifelse(is.na(var.max), var.mean, var.max)
}
```

Read in the raw data
```{r}
bfdata <- read_csv(here("raw-data/biteforce-data.csv"))
ecodata <- read_csv(here("raw-data/ecological-data.csv"))
```

## Data cleaning

```{r}
bfdata1 <- 
  bfdata %>%  
  # Remove low quality data
  filter(DataQuality != "Remove") %>%
  # Remove juveniles and sub-adults
  filter(!Age %in% c("JV", "SA")) %>% 
  # Remove unneeded columns
  dplyr::select(-c(Journal, Year, BinomialStudy, SC, Notes, 
            BM.MIN, SVL.MIN, HL.MIN, HW.MIN, HH.MIN, 
            LJL.MIN, BF.MIN, Age)) %>%
  # Replace NA values with "unknown" in sex column
  mutate(Sex = replace_na(Sex, "unknown")) %>% 
  # Append higher taxonomy column
  mutate(HigherTaxonomy = 
           case_when(Family == "Gekkonidae" | Family == "Phyllodactylidae" | 
                     Family == "Carphodactylidae" | Family == "Diplodactylidae" | 
                     Family == "Eublepharidae" | Family == "Sphaerodactylidae" | 
                     Family == "Pygopodidae" ~ "Gekkota",
                     
                     Family == "Cordylidae" | Family == "Scincidae" |
                     Family == "Gerrhosauridae" | 
                     Family == "Xantusiidae" ~ "Scincoidea", 
                
                     Family == "Anguidae" | Family == "Diploglossidae" | 
                     Family == "Shinisauridae" | Family == "Xenosauridae" |
                     Family == "Helodermatidae" | Family == "Lanthanotidae" | 
                     Family == "Varanidae" ~ "Anguimorpha",  
                     
                     Family == "Dibamidae" ~ "Dibamia",
                     
                     Family == "Chamaeleonidae" | Family == "Agamidae" 
                     ~ "Acrodonta",
                    
                     Family == "Phrynosomatidae" | Family == "Iguanidae" | 
                     Family == "Crotaphytidae" | Family == "Dactyloidae" |
                     Family == "Tropiduridae" | Family == "Liolaemidae" | 
                     Family == "Corytophanidae" | Family == "Hoplocercidae" | 
                     Family == "Leiocephalidae" | Family == "Leiosauridae" | 
                     Family == "Opluridae" | Family == "Polychrotidae" 
                     ~ "Pleurodonta",
                     
                     Family == "Alopoglossidae" | Family == "Gymnophthalmidae" | 
                     Family == "Teiidae" |
                     Family == "Amphisbaenidae" | Family == "Bipedidae" | 
                     Family == "Blanidae" | Family == "Cadeidae" | 
                     Family == "Rhineuridae" | Family == "Trogonophidae" |
                     Family == "Lacertidae" 
                     ~ "Lacertoidea",
      
                     Family == "Sphenodontidae" ~ "Rhynchocephalia")) %>%
  # Replace spaces with underscores (_) so they match up with species names in the tree
  mutate(tips = str_replace(BinomialReptileDatabase, " ", "_")) %>%
  # Convert SE to SD where we have SE but no SD 
  # Note that for BF we use SSBF (sample size for bite force)
  # For morphological variables we use SSM (sample size morphometrics)
  mutate(BF.SD = case_when(is.na(BF.SD) & !is.na(BF.SE) ~ se2sd(BF.SE, SSBF),
                           TRUE ~ BF.SD)) %>%
  mutate(SVL.SD = case_when(is.na(SVL.SD) & !is.na(SVL.SE) ~ se2sd(SVL.SE, SSM),
                           TRUE ~ SVL.SD)) %>%
  mutate(BM.SD = case_when(is.na(BM.SD) & !is.na(BM.SE) ~ se2sd(BM.SE, SSM),
                           TRUE ~ BM.SD)) %>%
  mutate(HL.SD = case_when(is.na(HL.SD) & !is.na(HL.SE) ~ se2sd(HL.SE, SSM),
                           TRUE ~ HL.SD)) %>%
  mutate(HW.SD = case_when(is.na(HW.SD) & !is.na(HW.SE) ~ se2sd(HW.SE, SSM),
                           TRUE ~ HW.SD)) %>%
  mutate(HH.SD = case_when(is.na(HH.SD) & !is.na(HH.SE) ~ se2sd(HH.SE, SSM),
                           TRUE ~ HH.SD)) %>%
  mutate(LJL.SD = case_when(is.na(LJL.SD) & !is.na(LJL.SE) ~ se2sd(LJL.SE, SSM),
                           TRUE ~ LJL.SD)) %>%
# Make acrodont variable
  mutate(acrodont = case_when(HigherTaxonomy == "Acrodonta" |
                              HigherTaxonomy == "Rhynchocephalia" ~ "Acrodont",
                              HigherTaxonomy != "Acrodonta" &
                              HigherTaxonomy != "Rhynchocephalia" ~ "Non acrodont")) %>%
  # Add a column that says how many NAs each row has
  mutate(number_NAs = rowSums(is.na(.)))

# Create columns of BF and each morphological variable + the SD (as a proxy for the max)
# Note we use sum, not + so that if there is no SD this outputs the mean.
bfdata1 <-
  bfdata1 %>%
  rowwise() %>% 
  mutate(bf_plus_sd = sum(BF, BF.SD, na.rm = TRUE)) %>%
  mutate(svl_plus_sd = sum(SVL, SVL.SD, na.rm = TRUE)) %>%
  mutate(bm_plus_sd = sum(BM, BM.SD, na.rm = TRUE)) %>%
  mutate(hl_plus_sd = sum(HL, HL.SD, na.rm = TRUE)) %>%
  mutate(hw_plus_sd = sum(HW, HW.SD, na.rm = TRUE)) %>%
  mutate(hh_plus_sd = sum(HH, HH.SD, na.rm = TRUE)) %>%
  mutate(ljl_plus_sd = sum(LJL, LJL.SD, na.rm = TRUE))

# This introduces 0 where there is no value for the variable at all
# Replace with NAs
  bfdata1[bfdata1 == 0] <- NA

# look at it
# bfdata1
  
# Get a few summary numbers  
# Total number of records
dim(bfdata1)
# How many records have max BF?
bfdata1 %>%
filter(!is.na(BF.MAX)) %>%
  group_by(Sex) %>%
  summarise(n())
# How many species have max BF?           
bfdata1 %>%
filter(!is.na(BF.MAX)) %>%
  dplyr::select(BinomialReptileDatabase) %>%
  distinct()
# How many studies have max BF?           
bfdata1 %>%
filter(!is.na(BF.MAX)) %>%
  dplyr::select(Citation) %>%
  distinct()
```

Expand the ecological data to get more useful habitat categories
```{r}
ecodata1 <- 
  ecodata %>%
  mutate(arboreal = case_when(Lifestyle == "Arboreal"  ~ "Arboreal", 
                              Lifestyle != "Arboreal" ~ "Other")) %>%
  mutate(fossorial = case_when(Lifestyle == "Fossorial"  ~ "Fossorial", 
                               Lifestyle != "Fossorial" ~ "Other")) %>%
  mutate(saxicolous = case_when(Lifestyle == "Arboreal/Saxicolous"  | Lifestyle == "Arboreal/Terrestrial/Saxicolous" |
                                Lifestyle == "Saxicolous" | Lifestyle == "Terrestrial/Saxicolous"  ~ "Saxicolous", 
                                Lifestyle != "Arboreal/Saxicolous"  & Lifestyle != "Arboreal/Terrestrial/Saxicolous" &
                                Lifestyle != "Saxicolous" & Lifestyle != "Terrestrial/Saxicolous"  ~ "Other"))
```

Combine the two datasets
```{r}
bfdata_all <- 
  ecodata1 %>% 
  inner_join(bfdata1, by = "BinomialReptileDatabase")
```

## Create new datasets for each relevant subset of the data

### Maximum BF or mean BF only
1. max_bf:  maximum bite force for each species (males and females combined). 
2. max_bf_males: maximum bite force for each species (males only)
3. max_bf_females: maximum bite force for each species (females only)

### Maximum BF or mean BF + SD.
4. avg_bf_plus_sd:  average bite force + standard deviation for each species (overall = males and females combined)
5. avg_bf_plus_sd_males: average bite force + standard deviation for each species (males only)
6. avg_bf_plus_sd_females: average bite force + standard deviation for each species (females only)

### 1. max_bf_overall 
We create new columns for BF and all the morph measurements (BM,SVL,HH,HW,HL LJL) and use a function (get_max) which returns either the max value recorded for that row, or the mean value for that row if there is no maximum available. Most papers report either mean OR max for these variables.

Next we select the row with the highest biteforce for each species. For instances where two individuals of the same species have the same max BF (duplicates) we keep the individual with the fewest NAs (i.e. the better quality data). We also keep the morphological data associated with that row.

```{r}  
max_bf <- 
   bfdata_all %>% 
   # Extract the max or mean for each individual
   mutate(max_bf = get_max(BF.MAX, BF))  %>%
   mutate(max_bm = get_max(BM.MAX, BM))  %>%
   mutate(max_svl = get_max(SVL.MAX, SVL))  %>%  
   mutate(max_hl = get_max(HL.MAX, HL))  %>%
   mutate(max_hw = get_max(HW.MAX, HW))  %>%
   mutate(max_hh = get_max(HH.MAX, HH))  %>%
   mutate(max_ljl = get_max(LJL.MAX, LJL))  %>%
  
   # Filter by the max bite force for each species
   group_by(BinomialReptileDatabase) %>%
   filter(max_bf == max(max_bf)) %>%
  
   # Where two individuals in a species have the same max, filter by data quality
   arrange(number_NAs) %>% 
   distinct(BinomialReptileDatabase, .keep_all = TRUE) %>%
  
   # Select columns we want
   dplyr::select(BinomialReptileDatabase, Sex, Family, HigherTaxonomy, tips,
          SSM, SSBF, max_bf, max_bm, max_svl, max_hl, max_hw, max_hh, max_ljl,
          arboreal, fossorial, Diet, acrodont, saxicolous)

# Look at it
max_bf
```  

### 2. max_bf_males
Exact same steps as max_bf, except filtering for males and removing species with unknown sex.

```{r}
max_bf_males <- 
   bfdata_all %>% 
   # Filter only males and remove species with unknown sex
   filter(Sex == "Male" & Sex != "unknown") %>%
   # Extract the max or mean for each individual
   mutate(max_bf = get_max(BF.MAX, BF))  %>%
   mutate(max_bm = get_max(BM.MAX, BM))  %>%
   mutate(max_svl = get_max(SVL.MAX, SVL))  %>%  
   mutate(max_hl = get_max(HL.MAX, HL))  %>%
   mutate(max_hw = get_max(HW.MAX, HW))  %>%
   mutate(max_hh = get_max(HH.MAX, HH))  %>%
   mutate(max_ljl = get_max(LJL.MAX, LJL))  %>%
  
   # Filter by the max bite force for each species
   group_by(BinomialReptileDatabase) %>%
   filter(max_bf == max(max_bf)) %>%
  
   # Where two individuals in a species have the same max, filter by data quality
   arrange(number_NAs) %>% 
   distinct(BinomialReptileDatabase, .keep_all = TRUE) %>%
  
   # Select columns we want
   dplyr::select(BinomialReptileDatabase, Sex, Family, HigherTaxonomy,  tips,
          SSM, SSBF, max_bf, max_bm, max_svl, max_hl, max_hw, max_hh, max_ljl,
          arboreal, fossorial, Diet, acrodont, saxicolous)
```    

### 3. max_bf_females
Exact same steps as max_bf, except filtering for females only and removing unknown sex species

```{r}
max_bf_females <- 
    bfdata_all %>% 
    # Filter only females and remove species with unknown sex:
    filter(Sex == "Female" & Sex != "unknown") %>%
    
   # Extract the max or mean for each individual
   mutate(max_bf = get_max(BF.MAX, BF))  %>%
   mutate(max_bm = get_max(BM.MAX, BM))  %>%
   mutate(max_svl = get_max(SVL.MAX, SVL))  %>%  
   mutate(max_hl = get_max(HL.MAX, HL))  %>%
   mutate(max_hw = get_max(HW.MAX, HW))  %>%
   mutate(max_hh = get_max(HH.MAX, HH))  %>%
   mutate(max_ljl = get_max(LJL.MAX, LJL))  %>%
  
   # Filter by the max bite force for each species
   group_by(BinomialReptileDatabase) %>%
   filter(max_bf == max(max_bf)) %>%
  
   # Where two individuals in a species have the same max, filter by data quality
   arrange(number_NAs) %>% 
   distinct(BinomialReptileDatabase, .keep_all = TRUE) %>%
  
   # Select columns we want
   dplyr::select(BinomialReptileDatabase, Sex, Family, HigherTaxonomy,  tips,
          SSM, SSBF, max_bf, max_bm, max_svl, max_hl, max_hw, max_hh, max_ljl,
          arboreal, fossorial, Diet, acrodont, saxicolous)
```

### 4.  max_bf_sd
The code is very similar to max_bf, the only difference is if we don't have a max this time we use mean + SD not mean. If the row does not have SD, then the mean alone is used.
 
```{r}    
max_bf_sd <- 
  bfdata_all %>%
  # Extract the max or mean+SD for each individual
   mutate(max_bf = get_max(BF.MAX, bf_plus_sd))  %>%
   mutate(max_bm = get_max(BM.MAX, bm_plus_sd))  %>%
   mutate(max_svl = get_max(SVL.MAX, svl_plus_sd))  %>%  
   mutate(max_hl = get_max(HL.MAX, hl_plus_sd))  %>%
   mutate(max_hw = get_max(HW.MAX, hw_plus_sd))  %>%
   mutate(max_hh = get_max(HH.MAX, hh_plus_sd))  %>%
   mutate(max_ljl = get_max(LJL.MAX, ljl_plus_sd))  %>%
   # Filter by the max bite force + sd for each species
   group_by(BinomialReptileDatabase) %>%
   filter(max_bf == max(max_bf)) %>%
   # Where two individuals in a species have the same max, filter by data quality
   arrange(number_NAs) %>% 
   distinct(BinomialReptileDatabase, .keep_all = TRUE) %>%
   # Select columns we want
    dplyr::select(BinomialReptileDatabase, Sex, Family, HigherTaxonomy,  tips,
           SSM, SSBF, max_bf, max_bm, max_svl, max_hl, max_hw, max_hh, max_ljl,
          arboreal, fossorial, Diet, acrodont, saxicolous)

# Look at it
max_bf_sd 
# How many of each sex?
max_bf_sd %>% group_by(Sex) %>% summarise(n())
```  

### 5. max_bf_sd_males

```{r}
max_bf_sd_males <- 
  bfdata_all %>%
  # Extract the max or mean+SD for each individual
   mutate(max_bf = get_max(BF.MAX, bf_plus_sd))  %>%
   mutate(max_bm = get_max(BM.MAX, bm_plus_sd))  %>%
   mutate(max_svl = get_max(SVL.MAX, svl_plus_sd))  %>%  
   mutate(max_hl = get_max(HL.MAX, hl_plus_sd))  %>%
   mutate(max_hw = get_max(HW.MAX, hw_plus_sd))  %>%
   mutate(max_hh = get_max(HH.MAX, hh_plus_sd))  %>%
   mutate(max_ljl = get_max(LJL.MAX, ljl_plus_sd))  %>%
  # Filter for males only and remove species with unknown sex 
  filter(Sex == "Male" & Sex != "unknown") %>%
  # Filter by the max bite force + sd for each species
   group_by(BinomialReptileDatabase) %>%
   filter(max_bf == max(max_bf)) %>%
   # Where two individuals in a species have the same max, filter by data quality
   arrange(number_NAs) %>% 
   distinct(BinomialReptileDatabase, .keep_all = TRUE) %>%
   # Select columns we want
    dplyr::select(BinomialReptileDatabase, Sex, Family, HigherTaxonomy, tips,
           SSM, SSBF, max_bf, max_bm, max_svl, max_hl, max_hw, max_hh, max_ljl,
          arboreal, fossorial, Diet, acrodont, saxicolous)
```

### 6. max_bf_sd_females

```{r}
max_bf_sd_females <- 
  bfdata_all %>%
  # Extract the max or mean+SD for each individual
   mutate(max_bf = get_max(BF.MAX, bf_plus_sd))  %>%
   mutate(max_bm = get_max(BM.MAX, bm_plus_sd))  %>%
   mutate(max_svl = get_max(SVL.MAX, svl_plus_sd))  %>%  
   mutate(max_hl = get_max(HL.MAX, hl_plus_sd))  %>%
   mutate(max_hw = get_max(HW.MAX, hw_plus_sd))  %>%
   mutate(max_hh = get_max(HH.MAX, hh_plus_sd))  %>%
   mutate(max_ljl = get_max(LJL.MAX, ljl_plus_sd))  %>%
  # Filter for males only and remove species with unknown sex 
  filter(Sex == "Female" & Sex != "unknown") %>%
  # Filter by the max bite force + sd for each species
   group_by(BinomialReptileDatabase) %>%
   filter(max_bf == max(max_bf)) %>%
   # Where two individuals in a species have the same max, filter by data quality
   arrange(number_NAs) %>% 
   distinct(BinomialReptileDatabase, .keep_all = TRUE) %>%
   # Select columns we want
    dplyr::select(BinomialReptileDatabase, Sex, Family, HigherTaxonomy, tips,
           SSM, SSBF, max_bf, max_bm, max_svl, max_hl, max_hw, max_hh, max_ljl,
          arboreal, fossorial, Diet, acrodont, saxicolous)
```
  
Write each dataset to csv for further analyses
```{r}
write_csv(bfdata_all, file = here("data/biteforce-ecology-data.csv"))
write_csv(max_bf, file = here("data/max_bf.csv"))
write_csv(max_bf_males, file = here("data/max_bf_males.csv"))
write_csv(max_bf_females, file = here("data/max_bf_females.csv"))
write_csv(max_bf_sd, file = here("data/max_bf_sd.csv"))
write_csv(max_bf_sd_males, file = here("data/max_bf_sd_males.csv"))
write_csv(max_bf_sd_females, file = here("data/max_bf_sd_females.csv"))
```
