==> analyses/03C-analyses-mean-basictree.Rmd <==
---
title: "PGLS analyses - using original tree"
output: html_document
---

# PGLS analyses and graphs

## Preparation

Load libraries
```{r, message = FALSE, warning = FALSE}
library(ape)
library(caper)
library(tidyverse)
library(patchwork)
library(broom)
library(here)
```

Read in the data and ensure these are dataframes.
```{r, message = FALSE}
max_bf <- read_csv(here("data/max_bf.csv"))
max_bf <- data.frame(max_bf)

max_bf_females <- read_csv(here("data/max_bf_females.csv"))
max_bf_females <- data.frame(max_bf_females)

max_bf_males <- read_csv(here("data/max_bf_males.csv"))
max_bf_males <- data.frame(max_bf_males)
```

Read in phylogeny
```{r}
tree <- read.nexus(here("data/lepidosaur-tree.nex"))
```

Create comparative data objects with phylogeny and data combined

```{r}
lepidosaur <- comparative.data(phy = tree, data = max_bf, 
                               names.col = tips, vcv = TRUE,
                               warn.dropped = FALSE, na.omit = FALSE)

# Check
# str(lepidosaur)

# Females only
lepidosaur_females <- comparative.data(phy = tree, data = max_bf_females, 
                               names.col = tips, vcv = TRUE,
                               warn.dropped = FALSE, na.omit = FALSE)

# Check
# str(lepidosaur_females)

# Males only
lepidosaur_males <- comparative.data(phy = tree, data = max_bf_males, 
                               names.col = tips, vcv = TRUE,
                               warn.dropped = FALSE, na.omit = FALSE)

# Check
# str(lepidosaur_males)
```

Set up output files
```{r}
output <- data.frame(array(dim = c(64, 20)))
names(output) <- c("dataset", "sex", "tree", "morph", "covariate", "lambda", "residual_df", 
                   "df_morph", "F_morph", "p_morph", 
                   "df_covariate", "F_covariate", "p_covariate", 
                   "df_morphxcovariate", "F_morphxcovariate", "p_morphxcovariate", 
                   "adj.r2", "bonf_p_morph", "bonf_p_covariate", 
                   "bonf_p_morphxcovariate")

# Short function to help write output files
write_output <- function(model, dataset, sex, tree, morph, covariate, n){
  
  x <- tidy(anova(model))
  y <- summary(model)
  
  output[n, "dataset"] <- dataset  
  output[n, "sex"] <- sex  
  output[n, "tree"] <- tree 
  output[n, "morph"] <- morph
  output[n, "covariate"] <- covariate
  output[n, "lambda"] <- y$param[[2]]
  output[n, "residual_df"] <- max(x$df)
  output[n, "df_morph"] <- x$df[1]
  output[n, "F_morph"] <- x$statistic[1]
  output[n, "p_morph"] <- x$p.value[1]
  output[n, "df_covariate"] <- x$df[2]
  output[n, "F_covariate"] <- x$statistic[2]
  output[n, "p_covariate"] <- x$p.value[2]
  output[n, "df_morphxcovariate"] <- x$df[3]
  output[n, "F_morphxcovariate"] <- x$statistic[3]
  output[n, "p_morphxcovariate"] <- x$p.value[3]
  output[n, "adj.r2"] <- y$adj.r.squared
  return(output)
}
```

Run the models!!!

### SVL + DIET

```{r}
# Fit model
model1 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(Diet), data = lepidosaur,
                   lambda = "ML")

# Check model diagnostics
plot(model1)

# Look at outputs
anova(model1)
summary(model1)

# Save outputs
output <- write_output(model1, dataset = "mean", sex = "all", tree = "basic",
                       morph = "SVL", covariate = "diet", n = 1)
```

```{r}
# Fit model
model2 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(Diet), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model2)

# Look at outputs
anova(model2)
summary(model2)

# Save outputs
output <- write_output(model2, dataset = "mean", sex = "female", tree = "basic",
                       morph = "SVL", covariate = "diet", n = 2)
```

```{r}
# Fit model
model3 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(Diet), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model3)

# Look at outputs
anova(model3)
summary(model3)

# Save outputs
output <- write_output(model3, dataset = "mean", sex = "male", tree = "basic",
                       morph = "SVL", covariate = "diet", n = 3)
```

### HW + DIET

```{r}
# Fit model
model4 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(Diet), data = lepidosaur,
                   lambda = "ML")

# Check model diagnostics
plot(model4)

# Look at outputs
anova(model4)
summary(model4)

# Save outputs
output <- write_output(model4, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HW", covariate = "diet", n = 4)
```

```{r}
# Fit model
model5 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(Diet), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model5)

# Look at outputs
anova(model5)
summary(model5)

# Save outputs
output <- write_output(model5, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HW", covariate = "diet", n = 5)
```

```{r}
# Fit model
model6 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(Diet), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model6)

# Look at outputs
anova(model6)
summary(model6)

# Save outputs
output <- write_output(model6, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HW", covariate = "diet", n = 6)
```

### SVL + ARBOREAL

```{r}
# Fit model
# Note had to raise lower bound of lambda slightly due to it getting stuck in ML space
# if I don't...
model7 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(arboreal), data = lepidosaur,
                   lambda = "ML", bounds = list(lambda = c(0.01, 1)))

# Check model diagnostics
plot(model7)

# Look at outputs
anova(model7)
summary(model7)

# Save outputs
output <- write_output(model7, dataset = "mean", sex = "all", tree = "basic",
                       morph = "SVL", covariate = "arboreal", n = 7)
```

```{r}
# Fit model
model8 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(arboreal), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model8)

# Look at outputs
anova(model8)
summary(model8)

# Save outputs
output <- write_output(model8, dataset = "mean", sex = "female", tree = "basic",
                       morph = "SVL", covariate = "arboreal", n = 8)
```

```{r}
# Fit model
model9 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(arboreal), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model9)

# Look at outputs
anova(model9)
summary(model9)

# Save outputs
output <- write_output(model9, dataset = "mean", sex = "male", tree = "basic",
                       morph = "SVL", covariate = "arboreal", n = 9)
```

### HW + ARBOREAL

```{r}
# Fit model
model10 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(arboreal), data = lepidosaur,
                   lambda = "ML")

# Check model diagnostics
plot(model10)

# Look at outputs
anova(model10)
summary(model10)

# Save outputs
output <- write_output(model10, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HW", covariate = "arboreal", n = 10)
```

```{r}
# Fit model
model11 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(arboreal), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model11)

# Look at outputs
anova(model11)
summary(model11)

# Save outputs
output <- write_output(model11, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HW", covariate = "arboreal", n = 11)
```

```{r}
# Fit model
model12 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(arboreal), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model12)

# Look at outputs
anova(model12)
summary(model12)

# Save outputs
output <- write_output(model12, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HW", covariate = "arboreal", n = 12)
```


### SVL + FOSSORIAL

```{r}
# Fit model
model13 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(fossorial), data = lepidosaur,
                   lambda = "ML")

# Check model diagnostics
plot(model13)

# Look at outputs
anova(model13)
summary(model13)

# Save outputs
output <- write_output(model13, dataset = "mean", sex = "all", tree = "basic",
                       morph = "SVL", covariate = "fossorial", n = 13)
```

> Cannot fit models to males and females as no sexed fossorial specimens exist

### HW + FOSSORIAL

```{r}
# Fit model
model14 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(fossorial), data = lepidosaur,
                   lambda = "ML")

# Check model diagnostics
plot(model14)

# Look at outputs
anova(model14)
summary(model14)

# Save outputs
output <- write_output(model14, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HW", covariate = "fossorial", n = 14)
```
> Cannot fit models to males and females as no sexed fossorial specimens exist

### SVL + HigherTaxonomy

Rhynchocephalia and Gekkota do not have enough species to fit the model so we need to remove them

```{r}
max_bf_x <- 
  max_bf %>%
  filter(HigherTaxonomy != "Rhynchocephalia" & HigherTaxonomy != "Gekkota"
         & HigherTaxonomy != "Anguimorpha") %>%
  data.frame()

lepidosaur_x <- comparative.data(phy = tree, data = max_bf_x, 
                               names.col = tips, vcv = TRUE,
                               warn.dropped = FALSE, na.omit = FALSE)

# females
max_bf_females_x <- 
  max_bf_females %>%
  filter(HigherTaxonomy != "Rhynchocephalia" & HigherTaxonomy != "Gekkota"
         & HigherTaxonomy != "Anguimorpha" & HigherTaxonomy != "Scincoidea") %>%
  data.frame()

lepidosaur_females_x <- comparative.data(phy = tree, data = max_bf_females_x, 
                               names.col = tips, vcv = TRUE,
                               warn.dropped = FALSE, na.omit = FALSE)

# males
max_bf_males_x <- 
  max_bf_males %>%
  filter(HigherTaxonomy != "Rhynchocephalia" & HigherTaxonomy != "Gekkota"&     HigherTaxonomy != "Anguimorpha" & HigherTaxonomy != "Scincoidea") %>%
  data.frame()

lepidosaur_males_x <- comparative.data(phy = tree, data = max_bf_males_x, 
                               names.col = tips, vcv = TRUE,
                               warn.dropped = FALSE, na.omit = FALSE)
```

```{r}
# Fit model
model15 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(HigherTaxonomy), 
                data = lepidosaur_x, lambda = "ML", bounds = list(lambda = c(0.1,1)))

# Check model diagnostics
plot(model15)

# Look at outputs
anova(model15)
summary(model15)

# Save outputs
output <- write_output(model15, dataset = "mean", sex = "all", tree = "basic",
                       morph = "SVL", covariate = "taxonomy", n = 15)
```

```{r}
# Fit model
model16 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(HigherTaxonomy), 
                data = lepidosaur_females_x, lambda = "ML")

# Check model diagnostics
plot(model16)

# Look at outputs
anova(model16)
summary(model16)

# Save outputs
output <- write_output(model16, dataset = "mean", sex = "female", tree = "basic",
                       morph = "SVL", covariate = "taxonomy", n = 16)
```

```{r}
# Fit model
model17 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(HigherTaxonomy), data = lepidosaur_males_x,
                   lambda = "ML", bounds = list(lambda = c(0.1,1)))

# Check model diagnostics
plot(model17)

# Look at outputs
anova(model17)
summary(model17)

# Save outputs
output <- write_output(model17, dataset = "mean", sex = "male", tree = "basic",
                       morph = "SVL", covariate = "taxonomy", n = 17)
```

### HW + taxonomy

```{r}
# Fit model
model18 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(HigherTaxonomy), data = lepidosaur_x,
                   lambda = "ML")

# Check model diagnostics
plot(model18)

# Look at outputs
anova(model18)
summary(model18)

# Save outputs
output <- write_output(model18, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HW", covariate = "taxonomy", n = 18)
```

```{r}
# Fit model
model19 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(HigherTaxonomy), data = lepidosaur_females_x,
                   lambda = "ML")

# Check model diagnostics
plot(model19)

# Look at outputs
anova(model19)
summary(model19)

# Save outputs
output <- write_output(model19, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HW", covariate = "taxonomy", n = 19)
```

```{r}
# Fit model
model20 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(HigherTaxonomy), data = lepidosaur_males_x,
                   lambda = "ML")

# Check model diagnostics
plot(model20)

# Look at outputs
anova(model20)
summary(model20)

# Save outputs
output <- write_output(model20, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HW", covariate = "taxonomy", n = 20)
```

## HH and HL measures

## HH + Diet

```{r}
# Fit model
model21 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(Diet), data = lepidosaur,
                   lambda = "ML")

# Check model diagnostics
plot(model21)

# Look at outputs
anova(model21)
summary(model21)

# Save outputs
output <- write_output(model21, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HH", covariate = "diet", n = 21)
```

```{r}
# Fit model
model22 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(Diet), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model22)

# Look at outputs
anova(model22)
summary(model22)

# Save outputs
output <- write_output(model22, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HH", covariate = "diet", n = 22)
```

```{r}
# Fit model
model23 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(Diet), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model23)

# Look at outputs
anova(model23)
summary(model23)

# Save outputs
output <- write_output(model23, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HH", covariate = "diet", n = 23)
```

### HL + DIET

```{r}
# Fit model
model24 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(Diet), data = lepidosaur,
                   lambda = "ML")

# Check model diagnostics
plot(model24)

# Look at outputs
anova(model24)
summary(model24)

# Save outputs
output <- write_output(model24, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HL", covariate = "diet", n = 24)
```

```{r}
# Fit model
model25 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(Diet), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model25)

# Look at outputs
anova(model25)
summary(model25)

# Save outputs
output <- write_output(model25, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HL", covariate = "diet", n = 25)
```

```{r}
# Fit model
model26 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(Diet), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model26)

# Look at outputs
anova(model26)
summary(model26)

# Save outputs
output <- write_output(model26, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HL", covariate = "diet", n = 26)
```

## HH + ARBOREAL

```{r}
# Fit model
model27 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(arboreal), data = lepidosaur,
                   lambda = "ML")

# Check model diagnostics
plot(model27)

# Look at outputs
anova(model27)
summary(model27)

# Save outputs
output <- write_output(model27, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HH", covariate = "arboreal", n = 27)
```

```{r}
# Fit model
model28 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(arboreal), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model28)

# Look at outputs
anova(model28)
summary(model28)

# Save outputs
output <- write_output(model28, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HH", covariate = "arboreal", n = 28)
```

```{r}
# Fit model
model29 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(arboreal), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model29)

# Look at outputs
anova(model29)
summary(model29)

# Save outputs
output <- write_output(model29, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HH", covariate = "arboreal", n = 29)
```

### HL + ARBOREAL

```{r}
# Fit model
model30 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(arboreal), data = lepidosaur,
                   lambda = "ML")

# Check model diagnostics
plot(model30)

# Look at outputs
anova(model30)
summary(model30)

# Save outputs
output <- write_output(model30, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HL", covariate = "arboreal", n = 30)
```

```{r}
# Fit model
model31 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(arboreal), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model31)

# Look at outputs
anova(model31)
summary(model31)

# Save outputs
output <- write_output(model31, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HL", covariate = "arboreal", n = 31)
```

```{r}
# Fit model
model32 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(arboreal), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model32)

# Look at outputs
anova(model32)
summary(model32)

# Save outputs
output <- write_output(model32, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HL", covariate = "arboreal", n = 32)
```

### HH + FOSSORIAL

```{r}
# Fit model
model33 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(fossorial), data = lepidosaur,
                   lambda = "ML")

# Check model diagnostics
plot(model33)

# Look at outputs
anova(model33)
summary(model33)

# Save outputs
output <- write_output(model33, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HH", covariate = "fossorial", n = 33)
```

> Cannot fit models to males and females as no sexed fossorial specimens exist

### HL + FOSSORIAL

```{r}
# Fit model
model34 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(fossorial), data = lepidosaur,
                   lambda = "ML")

# Check model diagnostics
plot(model34)

# Look at outputs
anova(model34)
summary(model34)

# Save outputs
output <- write_output(model34, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HL", covariate = "fossorial", n = 34)
```
> Cannot fit models to males and females as no sexed fossorial specimens exist

### HH + HigherTaxonomy

Rhynchocephalia and Gekkota do nto have enough species to fit the model so we need to remove them

```{r}
# Fit model
model35 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(HigherTaxonomy), 
                data = lepidosaur_x, lambda = "ML", bounds = list(lambda = c(0.1,1)))

# Check model diagnostics
plot(model35)

# Look at outputs
anova(model35)
summary(model35)

# Save outputs
output <- write_output(model35, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HH", covariate = "taxonomy", n = 35)
```

```{r}
# Fit model
model36 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(HigherTaxonomy), 
                data = lepidosaur_females_x, lambda = "ML")

# Check model diagnostics
plot(model36)

# Look at outputs
anova(model36)
summary(model36)

# Save outputs
output <- write_output(model36, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HH", covariate = "taxonomy", n = 36)
```

```{r}
# Fit model
model37 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(HigherTaxonomy), data = lepidosaur_males_x,
                   lambda = "ML", bounds = list(lambda = c(0.1,1)))

# Check model diagnostics
plot(model37)

# Look at outputs
anova(model37)
summary(model37)

# Save outputs
output <- write_output(model37, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HH", covariate = "taxonomy", n = 37)
```

### HL

```{r}
# Fit model
model38 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(HigherTaxonomy), data = lepidosaur_x,
                   lambda = "ML")

# Check model diagnostics
plot(model38)

# Look at outputs
anova(model38)
summary(model38)

# Save outputs
output <- write_output(model38, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HL", covariate = "taxonomy", n = 38)
```

```{r}
# Fit model
model39 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(HigherTaxonomy), data = lepidosaur_females_x,
                   lambda = "ML")

# Check model diagnostics
plot(model39)

# Look at outputs
anova(model39)
summary(model39)

# Save outputs
output <- write_output(model39, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HL", covariate = "taxonomy", n = 39)
```

```{r}
# Fit model
model40 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(HigherTaxonomy), data = lepidosaur_males_x,
                   lambda = "ML")

# Check model diagnostics
plot(model40)

# Look at outputs
anova(model40)
summary(model40)

# Save outputs
output <- write_output(model40, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HL", covariate = "taxonomy", n = 40)
```


### ACRODONTY

# SVL
```{r}
# Fit model
model41 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(acrodont), 
                data = lepidosaur, lambda = "ML")

# Check model diagnostics
plot(model41)

# Look at outputs
anova(model41)
summary(model41)

# Save outputs
output <- write_output(model41, dataset = "mean", sex = "all", tree = "basic",
                       morph = "SVL", covariate = "acrodont", n = 41)
```

```{r}
# Fit model
model42 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(acrodont), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model42)

# Look at outputs
anova(model42)
summary(model42)

# Save outputs
output <- write_output(model42, dataset = "mean", sex = "female", tree = "basic",
                       morph = "SVL", covariate = "acrodont", n = 42)
```

```{r}
# Fit model
model43 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(acrodont), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model43)

# Look at outputs
anova(model43)
summary(model43)

# Save outputs
output <- write_output(model43, dataset = "mean", sex = "male", tree = "basic",
                       morph = "SVL", covariate = "acrodont", n = 43)
```

# HW 
```{r}
# Fit model
model44 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(acrodont), 
                data = lepidosaur, lambda = "ML", bounds = list(lambda = c(0.1,1)))

# Check model diagnostics
plot(model44)

# Look at outputs
anova(model44)
summary(model44)

# Save outputs
output <- write_output(model44, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HW", covariate = "acrodont", n = 44)
```

```{r}
# Fit model
model45 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(acrodont), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model45)

# Look at outputs
anova(model45)
summary(model45)

# Save outputs
output <- write_output(model45, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HW", covariate = "acrodont", n = 45)
```

```{r}
# Fit model
model46 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(acrodont), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model46)

# Look at outputs
anova(model46)
summary(model46)

# Save outputs
output <- write_output(model46, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HW", covariate = "acrodont", n = 46)
```

# HH

```{r}
# Fit model
model47 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(acrodont), 
                data = lepidosaur, lambda = "ML")

# Check model diagnostics
plot(model47)

# Look at outputs
anova(model47)
summary(model47)

# Save outputs
output <- write_output(model47, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HH", covariate = "acrodont", n = 47)
```

```{r}
# Fit model
model48 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(acrodont), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model48)

# Look at outputs
anova(model48)
summary(model48)

# Save outputs
output <- write_output(model48, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HH", covariate = "acrodont", n = 48)
```

```{r}
# Fit model
model49 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(acrodont), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model49)

# Look at outputs
anova(model49)
summary(model49)

# Save outputs
output <- write_output(model49, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HH", covariate = "acrodont", n = 49)
```

# HL

```{r}
# Fit model
model50 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(acrodont), 
                data = lepidosaur, lambda = "ML")

# Check model diagnostics
plot(model50)

# Look at outputs
anova(model50)
summary(model50)

# Save outputs
output <- write_output(model50, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HL", covariate = "acrodont", n = 50)
```

```{r}
# Fit model
model51 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(acrodont), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model51)

# Look at outputs
anova(model51)
summary(model51)

# Save outputs
output <- write_output(model51, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HL", covariate = "acrodont", n = 51)
```

```{r}
# Fit model
model52 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(acrodont), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model52)

# Look at outputs
anova(model52)
summary(model52)

# Save outputs
output <- write_output(model52, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HL", covariate = "acrodont", n = 52)
```

# SAXICOLOUS

# SVL
```{r}
# Fit model
model53 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(saxicolous), 
                data = lepidosaur, lambda = "ML")

# Check model diagnostics
plot(model53)

# Look at outputs
anova(model53)
summary(model53)

# Save outputs
output <- write_output(model53, dataset = "mean", sex = "all", tree = "basic",
                       morph = "SVL", covariate = "saxicolous", n = 53)
```

```{r}
# Fit model
model54 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(saxicolous), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model54)

# Look at outputs
anova(model54)
summary(model54)

# Save outputs
output <- write_output(model54, dataset = "mean", sex = "female", tree = "basic",
                       morph = "SVL", covariate = "saxicolous", n = 54)
```

```{r}
# Fit model
model55 <- pgls(log(max_bf) ~ log(max_svl) * as.factor(saxicolous), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model55)

# Look at outputs
anova(model55)
summary(model55)

# Save outputs
output <- write_output(model55, dataset = "mean", sex = "male", tree = "basic",
                       morph = "SVL", covariate = "saxicolous", n = 55)
```

# HW 
```{r}
# Fit model
model56 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(saxicolous), 
                data = lepidosaur, lambda = "ML")

# Check model diagnostics
plot(model56)

# Look at outputs
anova(model56)
summary(model56)

# Save outputs
output <- write_output(model56, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HW", covariate = "saxicolous", n = 56)
```

```{r}
# Fit model
model57 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(saxicolous), data = lepidosaur_females,
                   lambda = "ML", bounds = list(lambda = c(0.1,1)))

# Check model diagnostics
plot(model57)

# Look at outputs
anova(model57)
summary(model57)

# Save outputs
output <- write_output(model57, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HW", covariate = "saxicolous", n = 57)
```

```{r}
# Fit model
model58 <- pgls(log(max_bf) ~ log(max_hw) * as.factor(saxicolous), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model58)

# Look at outputs
anova(model58)
summary(model58)

# Save outputs
output <- write_output(model58, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HW", covariate = "saxicolous", n = 58)
```

# HH

```{r}
# Fit model
model59 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(saxicolous), 
                data = lepidosaur, lambda = "ML")

# Check model diagnostics
plot(model59)

# Look at outputs
anova(model59)
summary(model59)

# Save outputs
output <- write_output(model59, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HH", covariate = "saxicolous", n = 59)
```

```{r}
# Fit model
model60 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(saxicolous), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model60)

# Look at outputs
anova(model60)
summary(model60)

# Save outputs
output <- write_output(model60, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HH", covariate = "saxicolous", n = 60)
```

```{r}
# Fit model
model61 <- pgls(log(max_bf) ~ log(max_hh) * as.factor(saxicolous), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model61)

# Look at outputs
anova(model61)
summary(model61)

# Save outputs
output <- write_output(model61, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HH", covariate = "saxicolous", n = 61)
```

# HL

```{r}
# Fit model
model62 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(saxicolous), 
                data = lepidosaur, lambda = "ML")

# Check model diagnostics
plot(model62)

# Look at outputs
anova(model62)
summary(model62)

# Save outputs
output <- write_output(model62, dataset = "mean", sex = "all", tree = "basic",
                       morph = "HL", covariate = "saxicolous", n = 62)
```

```{r}
# Fit model
model63 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(saxicolous), data = lepidosaur_females,
                   lambda = "ML")

# Check model diagnostics
plot(model63)

# Look at outputs
anova(model63)
summary(model63)

# Save outputs
output <- write_output(model63, dataset = "mean", sex = "female", tree = "basic",
                       morph = "HL", covariate = "saxicolous", n = 63)
```

```{r}
# Fit model
model64 <- pgls(log(max_bf) ~ log(max_hl) * as.factor(saxicolous), data = lepidosaur_males,
                   lambda = "ML")

# Check model diagnostics
plot(model64)

# Look at outputs
anova(model64)
summary(model64)

# Save outputs
output <- write_output(model64, dataset = "mean", sex = "male", tree = "basic",
                       morph = "HL", covariate = "saxicolous", n = 64)
```


### Bonferonni!!!
```{r}
output$bonf_p_morph <- p.adjust(output$p_morph)
output$bonf_p_covariate <- p.adjust(output$p_covariate)
output$bonf_p_morphxcovariate <- p.adjust(output$p_morphxcovariate)
```

Write all outputs to file

```{r}
#write_csv(output, file = here("outputs/pgls-results-basictree.csv"))
```
