==> analyses/03F-acrodont-sensitivity-analyses-mean.Rmd <==
---
title: "Biteforce body size and acrodonty sensitivity analyses"
output: html_document
---

Switches amphisbaenians to acrodont category and reruns the models.

## Preparation

Load libraries
```{r}
library(ape)
library(tidyverse)
library(knitr)
library(caper)
library(here)
```

Read in the data and ensure these are dataframes.
```{r, message = FALSE}
max_bf <- read_csv(here("data/max_bf.csv"))
max_bf <- data.frame(max_bf)
```

Make acrodont variable BUT make Trogonophidae acrodont too
```{r}
max_bf <-
  max_bf %>%
  mutate(acrodont = case_when(HigherTaxonomy == "Acrodonta" |
                              HigherTaxonomy == "Rhynchocephalia" |
                              Family == "Trogonophidae" ~ "Acrodont",
                              HigherTaxonomy != "Acrodonta" &
                              HigherTaxonomy != "Rhynchocephalia" &
                              Family != "Trogonophidae" ~ "Pleurodont"))
```

Read in phylogeny
```{r}
tree_plus <- read.nexus(here("data/lepidosaur-tree-plus.nex"))
```

Create comparative data objects with phylogeny and data combined
```{r}
lepidosaur <- comparative.data(phy = tree_plus, data = max_bf, 
                               names.col = tips, vcv = TRUE,
                               warn.dropped = FALSE, na.omit = FALSE)
```

Fit models to get regression lines
```{r}
# SVL
model1a <- pgls(log(max_bf) ~ log(max_svl)* as.factor(acrodont), data = lepidosaur,
                   lambda = "ML")

# HW
model2a <- pgls(log(max_bf) ~ log(max_hw)* as.factor(acrodont), data = lepidosaur,
                   lambda = "ML", bounds = list(lambda = c(0.1,1)))

# HL
model3a <- pgls(log(max_bf) ~ log(max_hl)* as.factor(acrodont), data = lepidosaur,
                   lambda = "ML")

# HH
model4a <- pgls(log(max_bf) ~ log(max_hh)* as.factor(acrodont), data = lepidosaur,
                   lambda = "ML")
```

Look at the results
```{r}
anova(model1a)
summary(model1a)

anova(model2a)
summary(model2a)

anova(model3a)
summary(model3a)

anova(model4a)
summary(model4a)
```
